+++ R/data.R
@@ -16,11 +16,11 @@
 #' @source \url{https://www.eia.gov/}
 "tickers_eia"
 
-#' CushingStocks
-#' @description EIA weekly Cushing crude stocks.
+#' eiaStocks
+#' @description EIA weekly crude, NG, ULSD and RBOB stocks.
 #' @format data frame
-#' @source \url{https://www.eia.gov/opendata/qb.php?sdid=PET.W_EPC0_SAX_YCUOK_MBBL.W}
-"CushingStocks"
+#' @source \url{https://www.eia.gov}
+"eiaStocks"
 
 #' cancrudeassays
 #' @description Data set with historical Canadian Crude Assays.
@@ -105,11 +105,6 @@
 #' @description USD IR Discount, Forward and Zero curves from RQuantlib::DiscountCurve - Parallel toy data set
 #' @format data frame
 "usSwapCurvesPar"
-
-#' ng_storage
-#' @description EIA NG Storage Data
-#' @format data frame
-"ng_storage"
 
 #' tradeCycle
 #' @description Crude Trading Trade Cycles