+++ NEWS.md
@@ -4,7 +4,7 @@
 + Summarised statistics dataset `cancrudeassayssum ` for Canadian Crude assays.
 + Crude oil qualities dataset `crudes ` from crudemonitor.ca and BP Assays.
 + Sample GIS data sets for North American `refineries` and `crudepipelines`.
-+ Added `CushingStocks` sample data set.
++ Added `eiaStocks` sample data set.
 + Upgrade for `dplyr 1.0.0`.
 + Sample data sets `dflong` and `dfwide` updated.